Clean up keywords
Removes a couple of node types that were not actually used.

Also inlines `case`, as it is superfluous inside a `caseDecl`.

I could have done a similar thing for `key`, `ref`, and `unique`, but
decided against it, since it may be more useful to have an actual
syntax node for these (even if their presence is already indicated by a
field).